Move `HeaderAvatarView` to its own file
